requirements.txt
django
gunicorn
scikit-learn
numpy
pillow
pandas
django-cors-headers